UC5
Added availableProducts. that checks if stock is higher then 0. and checks if the id is already on the list.

Added add item, that checks if the products is null, and if the id is equal or bigger then 1.

then the product exists and is able to be added.

then add the item to the grocerylistitems